+++ .gitignore
@@ -16,3 +16,4 @@ coverage.xml
 /coverage.xml
 .pytest_cache/
 .DS_Store
+*.sqlite3
