+++ README.md
@@ -6,7 +6,7 @@
 
 
 The drf-api-action Python package is designed to elevate your testing experience for Django Rest Framework (DRF) REST endpoints.
-With the api-action fixture/custom decorator api-action, this package empowers you to effortlessly test your REST endpoints as if they were conventional functions.
+With the api-action fixture, this package empowers you to effortlessly test your REST endpoints as if they were conventional functions.
 
 Features:
 
@@ -18,11 +18,6 @@ Features:
 
 * **Pagination Support**: Paginating easily through pages by a single kwarg.
 
-This package was designed in a way that you can use it in the following ways:
-
-* As a pytest fixture
-
-* As part of your ViewSet implementation
 
 ## Demo
 
@@ -46,7 +41,6 @@ pip install drf-api-action
 import pytest
 from tests.test_app.models import DummyModel
 from tests.test_app.views import DummyViewSetFixture
-from drf_api_action.fixtures import action_api
 ```
 
 #### Step 2: use the following action_api mark decorator:
@@ -56,17 +50,24 @@ from drf_api_action.fixtures import action_api
 e.g:
 our ViewSet is called `DummyViewSetFixture`
 ```python
+import pytest
+from tests.test_app.views import DummyViewSetFixture
+
 @pytest.mark.action_api(view_set_class=DummyViewSetFixture)
 def test_call_as_api_fixture(db, action_api):
     pass
 ```
 Now you can use all `DummyViewSetFixture` functionality!
 
-#### Step 3: write your test
+#### Step 3: write your tests
 
 e.g:
 our ViewSet is called `DummyViewSetFixture`
 ```python
+import pytest
+from tests.test_app.models import DummyModel
+from tests.test_app.views import DummyViewSetFixture
+
 @pytest.mark.action_api(view_set_class=DummyViewSetFixture)
 def test_call_as_api_fixture(db, action_api):
     dummy_model = DummyModel()
@@ -77,81 +78,13 @@ def test_call_as_api_fixture(db, action_api):
 
 ```
 
-### To use `drf-api-action` in your ViewSet, you need to follow these steps:
-
-#### Step 1: Import the Required Classes and Decorators
-
-Import the necessary classes and decorators from `drf-api-action` and `rest_framework`:
-
 ```python
-from drf_api_action.decorators import action_api
-from drf_api_action.mixins import APIRestMixin
-from rest_framework.viewsets import ModelViewSet
-```
+import pytest
+from tests.test_app.views import DummyViewSetFixture
 
-#### Step 2: Inherit `APIRestMixin` in your View 
-
-Create your view class by inheriting the `APIRestMixin` class.
-
-For example, we want to inherit `ModelViewSet` (Could be any ViewSet) in our View:
-
-```python
-class DummyView(APIRestMixin, ModelViewSet):
-    queryset = DummyModel.objects.all()
-    serializer_class = DummySerializer
-```
-
-Another example:
-
-```python
-class UsersViewSet(APIRestMixin, mixins.RetrieveModelMixin,
-                   mixins.ListModelMixin, viewsets.GenericViewSet):
-    serializer_class = UsersSerializer
-```
-
-#### Step 3: Define Your API Actions
-
-Use the `action_api` decorator instead of `action` decorator to define your API actions as functions inside your view class:
-
-From:
-```python
-    @action(detail=True, methods=["get"], serializer_class=DummySerializer)
-    def dummy(self, request, **kwargs):
-        serializer = self.get_serializer(instance=self.get_object())
-        return Response(data=serializer.data, status=status.HTTP_200_OK)
-```
-
-To:
-
-```python
-    @action_api(detail=True, methods=["get"], serializer_class=DummySerializer)
-    def dummy(self, request, **kwargs):
-        serializer = self.get_serializer(instance=self.get_object())
-        return Response(data=serializer.data, status=status.HTTP_200_OK)
-```
-
-In the example above, the `dummy` function is decorated with `action_api`.
-It specifies that the action requires a detail argument, supports the `GET` method, and uses the `DummySerializer` for serialization.
-
-#### Step 4: test REST methods
-
-* Create an instance of your view class and call the API actions as regular functions:
-
-```python
-def test_dummy():
-    api = DummyView()
-    result = api.dummy(pk=1)
-    assert result['dummy_int'] == 1
-```
-
-*query parameters/post payload are treated as function arguments as kwargs*
-
-
-* Exceptions are raised explicitly:
-```python
-def test_dummy():
-    api = DummyView()
-    result = api.dummy(pk='bbb')
+@pytest.mark.action_api(view_set_class=DummyViewSetFixture)
+def test_dummy(db, action_api):
+    result = action_api.dummy(pk='bbb')
     assert result['dummy_int'] == 1
 ```
 
@@ -187,14 +120,6 @@ filter_kwargs = {'pk': 'bb'}
 and so on....
 ```
 
-* Pagination support with `page`/`offset` kwarg (depends on `DEFAULT_PAGINATION_CLASS`):
-```python
->>> api = DummyAPIViewSet()
->>> response = api.by_dummy_int(request=None, dummy_int=1, page=2)
-
->>> {'count': 2, 'next': None, 'previous': '', 'results': [OrderedDict([('id', 2), ('dummy_int', 1)])]}
-
-```
 
 ## Package Testing
