+++ tests/test_app/views.py
@@ -5,8 +5,6 @@ from .models import DummyModel
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
-from drf_api_action.decorators import action_api
-from drf_api_action.mixins import APIRestMixin
 from rest_framework.serializers import IntegerField, ModelSerializer, Serializer
 
 
@@ -36,13 +34,13 @@ class DummyViewSet(ModelViewSet):
     queryset = DummyModel.objects.all()
     serializer_class = DummySerializer
 
-    @action_api(detail=True, methods=["get"], serializer_class=DummySerializer)
+    @action(detail=True, methods=["get"], serializer_class=DummySerializer)
     def api_dummy(self, request, **kwargs):
         serializer = self.get_serializer(instance=self.get_object())
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
-class DummyAPIViewSet(APIRestMixin, ModelViewSet):
+class DummyAPIViewSet(ModelViewSet):
     queryset = DummyModel.objects.all()
     serializer_class = DummySerializer
 
@@ -53,12 +51,12 @@ class DummyAPIViewSet(APIRestMixin, ModelViewSet):
             "pk": self.kwargs.get("pk"),
         }
 
-    @action_api(detail=True, methods=["get"], serializer_class=DummySerializer)
+    @action(detail=True, methods=["get"], serializer_class=DummySerializer)
     def api_dummy(self, request, **kwargs):
         serializer = self.get_serializer(instance=self.get_object())
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
-    @action_api(detail=False, methods=["post"], serializer_class=GetDummyByIntSerializer)
+    @action(detail=False, methods=["post"], serializer_class=GetDummyByIntSerializer)
     def by_dummy_int(self, request, **kwargs):
         self.get_serializer(data=request.data).is_valid(raise_exception=True)
         queryset = DummyModel.objects.filter(dummy_int=request.data["dummy_int"]).order_by("id")
